move exposed result datatypes to top level in obfuscation package
